requirements.txt
fastapi
uvicorn
numpy
opencv-contrib-python
mega.py